input new service

+++ Assets/_Application/Sources/App/Core/Ecs/Components/Player/UserCarInput.cs
@@ -4,7 +4,7 @@ namespace Sources.App.Core.Ecs.Components.Player
 {
     public struct UserCarInput : IComponent
     {
-        public int Vertical;
-        public int Horizontal;
+        public float Vertical;
+        public float Horizontal;
     }
 }
